added popups for pending load/offload, log + more
